Link down the page to dev instructions in readme
Found my self wanting to look at CONTRIBUTING.md when "development instructions" were mentioned in the readme because I had seen that file link go by already and the "Development" section of the readme was still below the fold. I thought a link down the page might add clarity but feel free to ignore this if you don't agree.